requirements.txt
# Kafka
confluent-kafka>=2.3.0,<3.0

# Web framework
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.29.0,<1.0

# ML
scikit-learn>=1.4.0,<2.0
joblib>=1.3.0,<2.0

# Graph
networkx>=3.2,<4.0

# LLM
openai>=1.12.0,<2.0

# Database
mysql-connector-python>=8.3.0,<9.0

# Config
pydantic-settings>=2.2.0,<3.0

# Utilities
python-dateutil>=2.9.0,<3.0
numpy>=1.26.0,<2.0
httpx>=0.27.0,<1.0

# Testing
pytest>=8.0.0,<9.0
